Rename def.rs to hir_db.rs
