+++ crates/apollo-compiler/src/database/mod.rs
@@ -2,12 +2,12 @@ pub mod db;
 pub mod hir;
 
 mod ast;
-mod def;
 mod document;
+mod hir_db;
 mod inputs;
 
 pub use ast::{AstDatabase, AstStorage};
 pub use db::RootDatabase;
-pub use def::{HirDatabase, HirStorage};
 pub use document::{DocumentDatabase, DocumentStorage};
+pub use hir_db::{HirDatabase, HirStorage};
 pub use inputs::{InputDatabase, InputStorage};
